refactor: 도메인과 같은 패키지구조로 변경 및 코드포맷팅
