+++ src/test/java/blackjack/domain/scoreboard/UserGameResultTest.java
@@ -23,7 +23,7 @@ class UserGameResultTest {
         );
 
         //when
-        UserGameResult userGameResult = new UserGameResult(resultCards, "유저",WinOrLose.WIN);
+        UserGameResult userGameResult = new UserGameResult(resultCards, "유저", WinOrLose.WIN);
 
         //then
         assertThat(userGameResult.getScore()).isEqualTo(20);
