+++ src/test/java/blackjack/domain/user/UsersTest.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class UsersTest {
     @DisplayName("유저 이름이 중복되면 에러가 검출되는지 확인")
     @Test
-    void whenDuplicateNames(){
+    void whenDuplicateNames() {
         List<User> duplicateUsers = Arrays.asList(new User("욘"), new User("웨지"), new User("포비"), new User("욘"));
 
         assertThatThrownBy(() -> new Users(duplicateUsers))
