+++ src/test/java/blackjack/domain/user/status/StatusTest.java
@@ -1,6 +1,5 @@
-package blackjack.domain.user;
+package blackjack.domain.user.status;
 
-import blackjack.domain.user.status.Status;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
